chore: rename module upsert to upsert_query
